Corrección de errores
- Falta hacer que funcione el cálculo de la velocidad

+++ p3/p6_IPC_Alvaro_Maria_Noe_Pablo/p6_IPC/src/medidas/TasaEfectividad.java
@@ -25,8 +25,8 @@ public class TasaEfectividad extends Medida {
         for (Interaccion i: completadas)
             if (objetivos.contains(i))
                 numObjCompletados++;
-
-        setMedida(numObjCompletados / objetivos.size());
+        
+        setMedida((double)numObjCompletados / (double)objetivos.size());
     }
     
 }
